requirements.txt
langchain
google-auth
google-auth-oauthlib
google-api-python-client
ollama
